# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: VeriHire — verified workforce management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /customers:
    get:
      operationId: listCustomers
      tags: [customers]
      responses:
        "200":
          description: Customers
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Customer" }
    post:
      operationId: createCustomer
      tags: [customers]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CustomerInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Customer" }
  /customers/{id}:
    get:
      operationId: getCustomer
      tags: [customers]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Customer
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Customer" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }

  # ---------- Bookings ----------
  /bookings:
    get:
      operationId: listBookings
      tags: [bookings]
      parameters:
        - in: query
          name: customerId
          schema: { type: integer }
        - in: query
          name: workerId
          schema: { type: integer }
        - in: query
          name: status
          schema:
            type: string
            enum: [active, completed, cancelled]
      responses:
        "200":
          description: Bookings
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Booking" }
    post:
      operationId: createBooking
      tags: [bookings]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BookingInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
  /bookings/{id}:
    get:
      operationId: getBooking
      tags: [bookings]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Booking
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
        "404":
          description: Not found
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /bookings/{id}/status:
    patch:
      operationId: updateBookingStatus
      tags: [bookings]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BookingStatusInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Booking" }
  /bookings/{id}/attendance:
    get:
      operationId: listAttendance
      tags: [attendance]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Attendance entries
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AttendanceEntry" }
    post:
      operationId: createAttendance
      tags: [attendance]
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AttendanceInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AttendanceEntry" }

  # ---------- Tasks ----------
...
  /reviews:
    get:
      operationId: listReviews
      tags: [reviews]
      parameters:
        - in: query
          name: workerId
          schema: { type: integer }
        - in: query
          name: customerId
          schema: { type: integer }
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Review" }
    post:
      operationId: createReview
      tags: [reviews]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReviewInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Review" }

  # ---------- Admin / Wow endpoints ----------
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error: { type: string }
      required: [error]

    # ---------- Worker ----------
    Worker:
      type: object
      properties:
        id: { type: integer }
        fullName: { type: string }
        phone: { type: string }
        photoUrl:
          type: ["string", "null"]
        category: { type: string, description: "maid | driver | guard | cook | nanny | gardener | other" }
        skills:
          type: array
          items: { type: string }
        bio:
          type: ["string", "null"]
        experienceYears: { type: integer }
        expectedMonthlySalary:
          type: ["integer", "null"]
        hourlyRate:
          type: ["integer", "null"]
        city: { type: string }
        serviceArea:
          type: ["string", "null"]
        cnicNumber: { type: string }
        cnicImageUrl:
          type: ["string", "null"]
        verificationStatus:
          type: string
          enum: [pending, approved, rejected]
        rejectionReason:
          type: ["string", "null"]
        averageRating: { type: number }
        reviewCount: { type: integer }
        completedJobs: { type: integer }
        availability:
          type: string
          enum: [available, busy, offline]
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required:
        - id
        - fullName
        - phone
        - category
        - skills
        - experienceYears
        - city
        - cnicNumber
        - verificationStatus
        - averageRating
        - reviewCount
        - completedJobs
        - availability
        - createdAt
        - updatedAt
...
    Customer:
      type: object
      properties:
        id: { type: integer }
        fullName: { type: string }
        phone: { type: string }
        city: { type: string }
        photoUrl: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
      required: [id, fullName, phone, city, createdAt]
    CustomerInput:
      type: object
      properties:
        fullName: { type: string }
        phone: { type: string }
        city: { type: string }
        photoUrl: { type: ["string", "null"] }
      required: [fullName, phone, city]

    # ---------- Booking (monthly hire) ----------
    Booking:
      type: object
      properties:
        id: { type: integer }
        customerId: { type: integer }
        workerId: { type: integer }
        startDate: { type: string, format: date-time }
        monthlySalary: { type: integer }
        status:
          type: string
          enum: [active, completed, cancelled]
        notes: { type: ["string", "null"] }
        worker: { $ref: "#/components/schemas/Worker" }
        customer: { $ref: "#/components/schemas/Customer" }
        attendanceSummary: { $ref: "#/components/schemas/AttendanceSummary" }
        createdAt: { type: string, format: date-time }
      required:
        - id
        - customerId
        - workerId
        - startDate
        - monthlySalary
        - status
        - createdAt
    BookingInput:
      type: object
      properties:
        customerId: { type: integer }
        workerId: { type: integer }
        startDate: { type: string, format: date-time }
        monthlySalary: { type: integer }
        notes: { type: ["string", "null"] }
      required: [customerId, workerId, startDate, monthlySalary]
    BookingStatusInput:
      type: object
      properties:
        status:
          type: string
          enum: [active, completed, cancelled]
      required: [status]
    AttendanceSummary:
      type: object
      properties:
        presentDays: { type: integer }
        absentDays: { type: integer }
        currentStreak: { type: integer }
      required: [presentDays, absentDays, currentStreak]
    AttendanceEntry:
      type: object
      properties:
        id: { type: integer }
        bookingId: { type: integer }
        date: { type: string, format: date }
        status:
          type: string
          enum: [present, absent]
        checkInTime: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
      required: [id, bookingId, date, status, createdAt]
    AttendanceInput:
      type: object
      properties:
        date: { type: string, format: date }
        status:
          type: string
          enum: [present, absent]
        checkInTime: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
      required: [date, status]

    # ---------- Task (on-demand) ----------
...
    Review:
      type: object
      properties:
        id: { type: integer }
        workerId: { type: integer }
        customerId: { type: integer }
        bookingId: { type: ["integer", "null"] }
        taskId: { type: ["integer", "null"] }
        rating: { type: integer }
        comment: { type: ["string", "null"] }
        customer: { $ref: "#/components/schemas/Customer" }
        createdAt: { type: string, format: date-time }
      required: [id, workerId, customerId, rating, createdAt]
    ReviewInput:
      type: object
      properties:
        workerId: { type: integer }
        customerId: { type: integer }
        bookingId: { type: ["integer", "null"] }
        taskId: { type: ["integer", "null"] }
        rating: { type: integer, minimum: 1, maximum: 5 }
        comment: { type: ["string", "null"] }
      required: [workerId, customerId, rating]

    # ---------- Admin ----------
...