# lib/api-zod/src/generated/types/attendanceEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VeriHire — verified workforce management API
 * OpenAPI spec version: 0.1.0
 */
import type { AttendanceEntryStatus } from "./attendanceEntryStatus";

export interface AttendanceEntry {
  id: number;
  bookingId: number;
  date: Date;
  status: AttendanceEntryStatus;
  /** @nullable */
  checkInTime?: string | null;
  /** @nullable */
  notes?: string | null;
  createdAt: Date;
}