# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VeriHire — verified workforce management API
 * OpenAPI spec version: 0.1.0
 */
import type { AttendanceSummary } from "./attendanceSummary";
import type { BookingStatus } from "./bookingStatus";
import type { Customer } from "./customer";
import type { Worker } from "./worker";

export interface Booking {
  id: number;
  customerId: number;
  workerId: number;
  startDate: Date;
  monthlySalary: number;
  status: BookingStatus;
  /** @nullable */
  notes?: string | null;
  worker?: Worker;
  customer?: Customer;
  attendanceSummary?: AttendanceSummary;
  createdAt: Date;
}