# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VeriHire — verified workforce management API
 * OpenAPI spec version: 0.1.0
 */

export interface BookingInput {
  customerId: number;
  workerId: number;
  startDate: Date;
  monthlySalary: number;
  /** @nullable */
  notes?: string | null;
}